jenkins: support autobuild key or release key
We need Jenkins to build ceph-deploy more often so we can have more
testing, but we don't want to sign these "unofficial" builds with the
official release key.

Conditionally select the release key or the autobuild key based upon the
$RELEASE environment variable. Jenkins will set this variable using a
parameter when it runs each ceph-deploy build. For formal releases, we
will set the parameter to "true".

Signed-off-by: Ken Dreyer <ken.dreyer@inktank.com>