Add better status message when adding new monitors
The `time.sleep(2)` in the `mon_add` method is not enough in some cases.
For example, when a low-spec slow VM acting as a monitor, it may still at
the state of 'probing' after the 2-seconds sleep. But in such cases the
`mon_status` method will just say the monitor *is not running*, it's kind of
confusion.

So maybe it would be better if provides user the current state of the
newly added monitor in the log message.

Signed-off-by: kfei <kfei@kfei.net>