write_conf() overwrite logic fixed for python 2.6
Explicitly set dict_type to OrderedDict in conf.ceph.parse(): it only
became the default for RawConfigParser in Python 2.7).

Parse and re-serialize the existing config file before comparing its
contents with expected configuration to make sure that the order of
options in the file doesn't affect the comparison.

Only overwrite the file if it's contents is different and --overwrite
option was used. If the contents matches the expected configuration,
leave it untouched.

conf.ceph.write_conf() has been superceded by host.remote.write_conf()
and is not called from anywhere, so it is removed.

Closes-Bug: #1333814