Modified the rpm_dist() and repository_url_part()
So they can use the newly defined distro.normalized_release object to determine
the correct OS version and return the information based on that. This should
work for all releases from RHEL, CentOS, and Scientific Linux 6 and greater.

Changed the test_centos.py file to use pytest fixtures to test multiple input
types for the various OSes we are testing for.

Signed-off-by: Tom Walsh <tom.walsh@expresshosting.net>