Return non-zero when gatherkeys fails
Immediately abort and exit when a needed key cannot be found
on the list of given hosts.

Define a new exception, KeyNotFoundError, and raise that from
gatherkeys() when the requested file does not exist on any of
the given hosts to try.  Since we now raise an exception on
error, there is no longer any need to return a value from
gatherkeys().

Fixes: 10043

Signed-off-by: Travis Rhoden <trhoden@redhat.com>