Add support for RHEL host module
When "use_rhceph" is set in cephdeploy.conf, no longer use
the centos host module as a stand-in for RHEL, use a new
rhel host module that is designed to be used for the Red Hat
distributed Red Hat Ceph product.

Right now this means not using Yum priorities, and installing
two new packages of ceph-mon and ceph-osd.  Eventually ceph-deploy
needs to be smarter of only installing ceph-mon or ceph-osd when a
host needs it, but for now just install both.

Signed-off-by: Travis Rhoden <trhoden@redhat.com>