[RM-12168] Don't use store_true and default=None
Not sure why this was ever seen as necessary.  With the store_true
action, it will be False if not present, True if passed on the CLI.

We just need to the boolean value, and there is no need to set
the flag to None instead of False when it is not present.

Fixes: #12168

Signed-off-by: Travis Rhoden <trhoden@redhat.com>