[fix-remove-ceph-create-keys-dep_cp] [RM-15451] Gather keys directly from the mon
Use the mon keyring to gather keys directly from the mon.
This is better than relying on:

  ceph-create-keys --cluster ${cluster_name} --id ${mon_id}

Which is started as a side effect of booting a mon. This side effect has
numerous issues as stated in this thread on ceph-devel:

    http://comments.gmane.org/gmane.comp.file-systems.ceph.devel/30552

The conclusion is that ceph will remove this side effect, after ceph-deploy
has a fix. This fix removes that dependency.

Signed-off-by: Owen Synge <osynge@suse.com>