[fix-remove-ceph-create-keys-dep_cp] docs.source.index: Update basic install
Updated the basic install instructions.

(1) Added a creating a new configuration section.
(2) Added reference to the 'new' subcommand
(3) Added reference to the 'mon' subcommand
(4) Improved Gather keys
(5) Added reference to the 'gatherkeys' subcommand
(6) Moved Admin hosts section as now it is more important.
(7) Added reference to the 'admin' subcommand

The gatherkeys subcommand does not depend upon ceph-create-keys so does not
have to have the side effect of making mons admin nodes. To get this in I
must add documentation to ceph-deploy.

Signed-off-by: Owen Synge <osynge@suse.com>