Fix UnboundLocalError when fail to get hostname
If we fail to get hostname, the `hosts.get` will raise an exception and
don't return any value, so the variable `distro` in [1] and [2] may be
undefined. What's worse, we may get previous value if we fail in second
time but succeed in first time.

In order to avoid the issues above, we should initialize the `distro`
value in each loop iteration.

[1] https://github.com/ceph/ceph-deploy/blob/master/ceph_deploy/mgr.py#L173
[2] https://github.com/ceph/ceph-deploy/blob/master/ceph_deploy/mds.py#L173